create course moved inside courses page
